change for waiting modem ack tested

+++ modem.h
@@ -29,8 +29,8 @@ unsigned char ModemStatus(void);
 unsigned char ModemSetSetings(void);
 unsigned char ModemSMSSend(unsigned char *number, unsigned char *text_sms);
 
-unsigned char ModemSendCom (const char *com_buff, unsigned int com_deluy_time_ms);
-unsigned char ModemSendData (const char *data_buff, unsigned int data_deluy_time_ms);
+unsigned char ModemSendCom (const char *com_buff, unsigned long com_deluy_time_ms);
+unsigned char ModemSendData (const char *data_buff, unsigned long data_deluy_time_ms);
 
 void ModemAnsverSet (const MODEM_ANSVER ansver) ;
